Enable -Wconversion for Linux builds
This adds -Wconversion (but not -Wsign-conversion) for compiling
cobalt on Linux.

This also adds missing explicit type conversions, and disable the
warning for the generated parser.

+++ cobalt/math/size_unittest.cc
@@ -14,7 +14,7 @@ namespace math {
 
 namespace {
 
-int TestSizeF(const SizeF& s) { return s.width(); }
+int TestSizeF(const SizeF& s) { return static_cast<int>(s.width()); }
 
 }  // namespace
